refactor(types): move View type
